feat: add outputs to api server/ingress ip
This adds outputs for the API server IP and the ingress IP. The recent
change to the GA ARO resource in the Azure RM provider required the use
of an input domain that is self-managed.  Because of this, you are required
to put in the API server and Ingress server as inputs in your DNS and
this will help facilitate automation surrounding that.

Signed-off-by: Dustin Scott <dscott@redhat.com>